+++ bpf/c/bpf_helpers.c
@@ -1,0 +1,1 @@
+../../syscall/samples/bpf/bpf_helpers.c