+++ syscall/samples/bpf/bpf_helpers.c
@@ -1,7 +1,7 @@
 #include <syscall.h>
 #include <sys/syscall.h>
 #include <unistd.h>
-#include "bpf.h"
+#include "bpf_helpers.h"
 
 int sys_bpf(int cmd, union bpf_attr *attr, unsigned int size)
 {