+++ bpf/c/foo.bpf.c
@@ -1,0 +1,9 @@
+int foo(void)
+{
+	/* bpf target default -O2, disable optimize */
+	volatile int a, b, c;
+	a = 1;
+	b = 2;
+	c = a + b;
+	return c;
+}