+++ bpf/libbpf/socket_helpers.h
@@ -1,4 +1,1 @@
-// SPDX-License-Identifier: (LGPL-2.1 OR BSD-2-Clause)
-#pragma once
-
-int open_raw_sock(const char *ifname);
+../../ipc/socket/socket_helpers.h