+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to
 
 #### Breaking Changes
 #### Added
+- bpf: add history.md;
 - qemu: add replay directory, symlink record.
 #### Changed
 #### Deprecated