+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to
 
 #### Breaking Changes
 #### Added
+- compiler: add jit/libgccjit/;
 - syscall/samples: add `bpf/bpf_helpers.c`;
 - bpf: add history.md, c/foo.bpf.c, kfunc/modules/`bpf_task_cwd_from_pid.c`, README.ISA.md, tcpdump.bpf.c;
 - qemu: add replay directory, symlink record.