+++ CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to
 - bpf: add history.md, c/foo.bpf.c, kfunc/modules/`bpf_task_cwd_from_pid.c`, README.ISA.md, tcpdump.bpf.c;
 - qemu: add replay directory, symlink record.
 #### Changed
+- `socket_helpers`: Rename `open_raw_sock` to `socket_raw`;
 #### Deprecated
 #### Removed
 - Remove `ipc/socket/raw/raw_helpers.c`, use `socket_helpers.c` instead;