+++ CHANGELOG.md
@@ -11,6 +11,19 @@ and this project adheres to
 
 #### Breaking Changes
 #### Added
+#### Changed
+#### Deprecated
+#### Removed
+#### Fixed
+#### Security
+#### Docs
+#### Tools
+
+
+## [v2.2.17] 2025-07-12
+
+#### Breaking Changes
+#### Added
 - compiler: add jit/libgccjit/;
 - syscall/samples: add `bpf/bpf_helpers.c`, bpf.bt;
 - bpf: add history.md, c/foo.bpf.c, kfunc/modules/`bpf_task_cwd_from_pid.c`, README.ISA.md, tcpdump.bpf.c;